cloudrun/genai/magix/backend/requirements.txt
# Copyright 2024 Google LLC
# 
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
# 
#     https://www.apache.org/licenses/LICENSE-2.0
# 
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

flask
flask-cors
vertexai
psycopg2-binary
python-dotenv
google-generativeai
openai
google-cloud-storage
cloud-sql-python-connector
cloud-sql-python-connector[pg8000]
SQLAlchemy
mistralai==1.0.1
pylint
pytest
pytest-mock
google-auth
google-auth-oauthlib
